requirements.txt
aiohttp>=3.8.0
asyncio
python-dotenv>=1.0.0
supabase>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.20
python-jose[cryptography]>=3.3.0
